Move new group method to separte group file in experiment module.
