+++ pyquil/tests/test_operator_estimation.py
@@ -21,7 +21,6 @@ from pyquil.experiment import (
     plusY,
     plusZ,
     zeros_state,
-    ExperimentResult,
 )
 from pyquil.gates import CNOT, CZ, H, I, RX, RY, RZ, X, Y
 from pyquil.operator_estimation import (
@@ -36,7 +35,6 @@ from pyquil.operator_estimation import (
     ratio_variance,
     _calibration_program,
     _pauli_to_product_state,
-    get_results_by_qubit_groups,
 )
 from pyquil.paulis import sI, sX, sY, sZ, PauliSum, PauliTerm
 from pyquil.quilbase import Pragma
@@ -2144,37 +2142,3 @@ def test_unitary_state_fidelity_readout_error(forest, use_seed):
     # how close is this state to |0>
     expected_fidelity = (np.cos(theta / 2)) ** 2
     np.testing.assert_allclose(expected_fidelity, estimated_fidelity, atol=2e-2)
-
-
-def test_results_by_qubit_groups():
-    er1 = ExperimentResult(
-        setting=ExperimentSetting(plusX(0), sZ(0)),
-        expectation=0.,
-        std_err=0.,
-        total_counts=1,
-    )
-
-    er2 = ExperimentResult(
-        setting=ExperimentSetting(plusX(0), sZ(1)),
-        expectation=0.,
-        std_err=0.,
-        total_counts=1,
-    )
-
-    er3 = ExperimentResult(
-        setting=ExperimentSetting(plusX(0), sX(0)*sZ(1)),
-        expectation=0.,
-        std_err=0.,
-        total_counts=1,
-    )
-
-    er4 = ExperimentResult(
-        setting=ExperimentSetting(plusX(0), sX(0)*sZ(2)),
-        expectation=0.,
-        std_err=0.,
-        total_counts=1,
-    )
-    groups = [(0,), (1,), (2, 0)]
-    res_by_group = get_results_by_qubit_groups([er1, er2, er3, er4], groups)
-
-    assert res_by_group == {(0,): [er1], (1,): [er2], (0, 2): [er1, er4]}
